refactor - split out commands into files
